requirements.txt
streamlit 
requests 
yfinance 
flask 
python-dotenv 
google-generativeai 
textblob 
yfinance 
flask-cors 
requests